requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot[ext]
firebase-admin
pydantic-settings
apscheduler
python-dotenv